An experiment in adding a client
I'd like to be able to talk to other huts and have serialization and
protocols be kind of invisible. This is a first attempt at that, with a
generic Client interface and a specific HttpClient implementation.

How things are encoded over the protocol is passed into the protocol
handler. Possibly that's not the right approach. In fact, I'm pretty
sure it's not.

Anyway, this isn't that important right now, so not committing it in a
branch and forgetting about it for a second.