The beginnings of a resource package
Resources are things that have unique identies in the system. They have
paths that are arrays that describe how to access them.

This is so far really simple as I'm not sure what the whole API will
look like. We need a way of building collections of paths and building
paths without being given an object.

Where do we store the path information? It's kind of in two places now
with Mux and resources.